# public/openapi.yaml
openapi: "3.0.3"
info:
  title: 智学AI - 开放API
  description: |
    智学AI教育平台开放API接口文档。提供AI辅导、LLM代理、题库管理、游戏化学习等接口。

    ## 快速开始
    1. **注册账号** - 在智学AI平台注册账号
    2. **获取API Key** - 在个人中心的「API Keys」页面创建虚拟API Key
    3. **调用API** - 使用 `X-API-Key` 请求头携带API Key发起请求

    ## 认证方式
    - **API Key** (`ApiKeyAuth`): 通过 `X-API-Key` 请求头或 `Authorization: Bearer <key>` 传递
    - **会话认证** (`BearerAuth`): 通过 NextAuth 的 session cookie 自动认证（仅限平台内调用）

    ## 错误响应
    所有接口在发生错误时都会返回统一格式：
    ```json
    { "error": "错误描述信息" }
    ```

    常见状态码：
    - `400` - 请求参数错误
    - `401` - 未认证 / API Key无效
    - `402` - 余额不足
    - `403` - 无权限
    - `404` - 资源不存在
    - `429` - 请求频率超限
    - `500` - 服务器内部错误

    ## 速率限制
    - 免费用户: 50 RPM
    - VIP用户: 200 RPM
    超出限制后返回 `429` 状态码。

  version: "1.0.0"
  contact:
    name: 智学AI开发团队
    email: dev@zixueai.com
    url: https://zixueai.com
servers:
  - url: "{baseUrl}"
    description: 智学AI API服务器
    variables:
      baseUrl:
        default: "https://api.zixueai.com"
        description: API服务器地址
        enum:
          - "https://api.zixueai.com"
          - "http://localhost:3000"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "从个人中心获取的API Key，格式: sk-edu-xxxxxx-xxxxxxxx"
    BearerAuth:
      type: http
      scheme: bearer
      description: "NextAuth Session Token（平台内自动携带）"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误描述信息
          example: "请先登录"
      required:
        - error
...
    UserProfileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "张三"
        email:
          type: string
          example: "zhangsan@example.com"
        avatar:
          type: string
        grade:
          type: string
          example: "初三"
        role:
          type: string
          example: "student"
        xp:
          type: integer
          example: 2500
        level:
          type: integer
          example: 8
        streak:
          type: integer
          example: 15
        maxStreak:
          type: integer
          example: 30
        balance:
          type: number
          example: 12.50
        badges:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              description:
                type: string
              icon:
                type: string
              earnedAt:
                type: string
                format: date-time
        stats:
          type: object
          properties:
            conversationsCount:
              type: integer
              example: 42
            errorNotesCount:
              type: integer
              example: 15
            notesCount:
              type: integer
              example: 8
            recordsCount:
              type: integer
              example: 120
            todayQuestionsDone:
              type: integer
              example: 25
            todayQuestionsCorrect:
              type: integer
              example: 20
            todayXPEarned:
              type: integer
              example: 150
            todayStudyDuration:
              type: integer
              description: 学习时长（分钟）
              example: 45
        hasApiKey:
          type: boolean
          example: true
    HomeStatsResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            avatar:
              type: string
            xp:
              type: integer
            level:
              type: integer
            streak:
              type: integer
            maxStreak:
              type: integer
        todayStats:
          type: object
          properties:
            duration:
              type: integer
              description: 学习时长（分钟）
            questionsDone:
              type: integer
            questionsCorrect:
              type: integer
            xpEarned:
              type: integer
            correctRate:
              type: integer
              description: 正确率百分比 0-100
        recentRecords:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              questionsDone:
                type: integer
              xpEarned:
                type: integer
              duration:
                type: integer
              createdAt:
                type: string
                format: date-time
...
paths:
...
  /api/user/profile:
    get:
      tags:
        - User
      summary: 获取用户资料
      description: 获取当前用户的完整资料，包含统计数据、坏章等信息
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 用户资料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
              example:
                id: "user_abc123"
                name: "张三"
                email: "zhangsan@example.com"
                avatar: null
                grade: "初三"
                role: "student"
                xp: 2500
                level: 8
                streak: 15
                maxStreak: 30
                balance: 12.50
                badges:
                  - id: "badge_001"
                    name: "初来乍到"
                    description: "完成首次学习"
                    icon: "star"
                    earnedAt: "2026-04-01T08:00:00Z"
                stats:
                  conversationsCount: 42
                  errorNotesCount: 15
                  notesCount: 8
                  recordsCount: 120
                  todayQuestionsDone: 25
                  todayQuestionsCorrect: 20
                  todayXPEarned: 150
                  todayStudyDuration: 45
                hasApiKey: true
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "用户不存在"
  /api/user/home-stats:
    get:
      tags:
        - User
      summary: 获取首页统计数据
      description: 获取首页所需的用户基本信息和今日学习统计数据
      operationId: getHomeStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 首页统计数据
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeStatsResponse"
              example:
                user:
                  id: "user_abc123"
                  name: "张三"
                  avatar: null
                  xp: 2500
                  level: 8
                  streak: 15
                  maxStreak: 30
                todayStats:
                  duration: 45
                  questionsDone: 25
                  questionsCorrect: 20
                  xpEarned: 150
                  correctRate: 80
                recentRecords:
                  - id: "rec_001"
                    questionsDone: 10
                    xpEarned: 60
                    duration: 20
                    createdAt: "2026-05-07T09:00:00Z"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "用户不存在"
...